Use protobuf messages (because the protocol needs a signed message instead of only the content of the message) - It's needed to show proof that some actor is malicious.

+++ dkg/keygen/round4.go
@@ -27,7 +27,7 @@ func (k *Keygen) r4Proceed() error {
 	}
 	k.Output = &LocalKeyShare{
 		Index:           k.PartyI,
-		Threshold:       uint16(len(k.Coefficients) - 1),
+		Threshold:       uint32(len(k.Coefficients) - 1),
 		ShareCount:      k.PartyCount,
 		PublicKey:       pk.Serialize(),
 		SecretShare:     k.skI.Serialize(),
